p4i integration Integrate 259339 from 4.3 to main: Compile in 64-bit mode. Fix casts from void * to int.
[git-p4: depot-paths = "//depot/qt/main/": change = 259541]

+++ tests/auto/qmetatype/tst_qmetatype.cpp
@@ -71,7 +71,7 @@ protected:
     void run()
     {
         for (int i = 0; i < 1000; ++i) {
-            const QByteArray name = QString("Bar%1_%2").arg(i).arg((int)QThread::currentThreadId()).toLatin1();
+            const QByteArray name = QString("Bar%1_%2").arg(i).arg((size_t)QThread::currentThreadId()).toLatin1();
             const char *nm = name.constData();
             int tp = qRegisterMetaType<Bar>(nm);
 #ifdef Q_OS_LINUX
